query3 anda. falta agregar el top 10
anda bien. en kaggle aparecen desordenados

+++ reviews_counter_accum/common/reviews_counter.py
@@ -17,7 +17,7 @@ class ReviewsCounter:
             score = review.score
             self.reviews[review.title] = (1, score)
         
-        if float(self.reviews[review.title][0]) > 4:
+        if int(self.reviews[review.title][0]) > 4:
             return review.title, self.reviews[review.title][0]
         return None, None
